Fix line-endings in index.html
Maybe this is a Mac thing, but my system is badly mangling the display
of these source files, owing to a wildly inconsistent use of end-of-line
characters.  I don't have any personal knowledge of the situation, but
I'm going to guess these files have suffered a long history of being
transfered again and again between machines with different EOL styles:

    * Unix, including modern macOS ("\n")
    * Windows ("\r\n")
    * Classic MacOS ("\r")

My personal preference is for Unix-style, but I like to think of myself
as a gracious guest.  The most recent edits appear to be Windows-style,
so I'm going to conform to that.  All this change does is ensure the
files are using the same line-ending throughout.  It helps readability,
but results in no functional change.