Convert journal two .au files to .wav
Unlike with journal one, some of the references to sound files are
embedded within the applets directly.  This requires modifying the
decompiled source, and then recompiling it back into .class files, so
that the applet still runs.  Analysis of the .class files reveals that
they were originally compiled using JDK 1.1, so I located a JDK of the
appropriate version and set up a VM that would let me run it.

The diff on the resulting .class files is rather significant for such a
small source code change.  I attribute this to the fact that the
decompiler could not perfectly reproduce the source file as it had been.